Use Shouldly and improved readability of tests

+++ src/BglReader/Elevation.cs
@@ -1,8 +1,8 @@
 ﻿namespace BglReader;
 
-public readonly record struct Elevation(float Value)
+public readonly record struct Elevation(double Value)
 {
-    public float Value { get; } = Value;
+    public double Value { get; } = Value;
 
     internal static Elevation FromBgl(int value) => new(value / 1000f);
 }
